code/requirements.txt
# Core dependencies
pynput>=1.7.6
Pillow>=9.0.0
python-docx>=0.8.11
psutil>=5.8.0

# GUI dependencies
customtkinter>=5.2.0

# Video export dependencies
moviepy>=1.0.3
opencv-python>=4.5.0

# AI dependencies
easyocr>=1.7.0
transformers>=4.20.0
torch>=1.12.0

# Optional dependencies for enhanced features
docx2pdf>=0.1.8
pywin32>=305; sys_platform == "win32"

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0 